Selection flickers when trying to change size of selection. https://bugs.webkit.org/show_bug.cgi?id=141404 rdar://problem/18824863
Reviewed by Benjamin Poulain.

When looking for the contracted range from the current range,
we were incorrectly choosing a selection whose rectangle is empty
as a best match candidate. This was throwing off all the logic
and producing a contracted range whose rectangle was bigger than the
expanded range, therefore producing a shrink threshold larger than the
growth one.

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::contractedRangeFromHandle):



Identifier: 159449@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@179847 268f45cc-cd09-0410-ab3c-d52691b4dbfc