Div having contentEditable and display:flex cannot be edited if it is empty. https://bugs.webkit.org/show_bug.cgi?id=141218
Reviewed by Ryosuke Niwa.

Source/WebCore:

RenderFlexibleBox should be treated as a candidate for visible position. Visible selection
in an editable area is recalculated whenever an inner value of div is changed. If the inner value
is empty, the visible selection recalculated with DeleteSelectionCommand::m_endingPostion
is not correct. Because RenderBlockFlow is only considered but not RenderFlexibleBox
in Position::isCandidate so that a calculated ending VisiblePosition for an editable div
having "display : flex" goes weird and VisibleSelection is empty accordingly.

Test: fast/events/key-events-in-editable-flexbox.html

* dom/Position.cpp:
(WebCore::Position::isCandidate):

LayoutTests:

This tests if an editable flexbox can be reedited after erasing all texts in it.

* fast/events/key-events-in-editable-flexbox-expected.txt: Added.
* fast/events/key-events-in-editable-flexbox.html: Added.


Identifier: 159500@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@179921 268f45cc-cd09-0410-ab3c-d52691b4dbfc