CachedResource::clearLoader() should self-destruct if nothing else retains the CachedResource. <https://webkit.org/b/141568> <rdar://problem/19800310>
Reviewed by Antti Koivisto.

Anything that may cause CachedResource::canDelete() to return true must also make sure
to call CachedResource::deleteIfPossible(), or we risk leaking the CachedResource.
This is because CachedResource employs an extremely convoluted lifetime mechanism that
depends on its presence in a number of collections, as well as internal counters and
state.

This is a speculative fix for a potential CachedResource leak that I'm not sure exists
in practice, but let's be good citizens here.

* loader/cache/CachedResource.cpp:
(WebCore::CachedResource::clearLoader):


Identifier: 159601@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@180068 268f45cc-cd09-0410-ab3c-d52691b4dbfc