Streamline unexported function build fixes https://bugs.webkit.org/show_bug.cgi?id=141761
Patch by Alexey Proskuryakov <ap@apple.com> on 2015-02-18
Reviewed by Alex Christensen.

Source/WebCore:

* Configurations/WebCore.unexp: Added some functions for symbols only used on newer
OS versions. Removed a special case for NodeList, to handle it uniformly with Node.

* Configurations/WebCore.xcconfig: Made the unexported list unconditional, because
it's not only Xcode 5 that is affected.

* bindings/scripts/CodeGeneratorJS.pm: (GenerateHeader): We already had a hack for
Node, NodeList needs an exactly the same one. Also updated the comments.

Tools:

* Scripts/check-for-weak-vtables-and-externals: (readXcode5SymbolsToIgnore): Deleted.
We don't need to ignore symbols that are not exported due to the .unexp file.



Identifier: 159782@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@180282 268f45cc-cd09-0410-ab3c-d52691b4dbfc